Adds on-filter-present functionality
Implements the `@QFOnFilterPresent` annotation to enable default values for certain filters when other filters are present in the request.

This feature allows for more dynamic and context-aware filtering, where the presence of one filter automatically activates default values for related filters, improving the usability and flexibility of the query filter.

Also fixes and improves default values and simplifies related code.