Adds default value support to QF annotations
Adds the ability to define default values and operations for QFElement, QFJsonElement, QFDiscriminator and QFCollectionElement annotations.
Also adds the possibility to define the order of the filters.

This change enables the definition of default filter values directly within the annotations, simplifying common use cases and reducing the need for external configuration.

Also includes changes to support OnFilterPresent annotation